Scripts/requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.7.1
bcrypt==3.2.2
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachelib==0.13.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==5.3.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
etelemetry==0.3.1
fastapi==0.115.8
filelock==3.17.0
fitz==0.0.1.dev2
Flask==3.1.0
Flask-Cors==3.0.10
Flask-Session==0.4.0
fonttools==4.55.6
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
gql==3.3.0
graphql-core==3.2.5
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.27.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kafka-python-ng==2.2.3
kiwisolver==1.4.8
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==0.3.1
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==2.2.2
openai==0.27.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
portalocker==2.10.1
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==3.4.0
setuptools==75.8.2
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
traits==7.0.2
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
xgboost==2.1.3
yarl==1.18.3